BC-4170 - move key down handling to `FileContentElement`
